HR System  for the recruitment process/recruitment_ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.0
chromadb==0.4.14
PyPDF2==3.0.1
python-docx==0.8.11
sqlalchemy==2.0.23
tenacity==8.2.3
PyJWT==2.11.0
python-multipart==0.0.6
streamlit==1.28.1
requests==2.31.0
scikit-learn==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
